requirements.txt
aiofiles==24.1.0
beartype==0.20.0
blinker==1.9.0
boto3==1.37.15
botocore==1.37.15
click==8.1.8
distro-info==1.0
Flask==3.1.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
priority==2.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Quart==0.20.0
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
